refactor(mobile): modular drawer asset selection lumen skeleton and cleanup
