+++ .changeset/calm-rivers-flow.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Use Lumen skeleton list in modular drawer asset selection, drop unused search focus handlers, align generic error horizontal margin
